interstitial refactor
