+++ node/src/components/contract_runtime.rs
@@ -76,7 +76,6 @@ use exec_queue::{ExecQueue, QueueItem};
 use metrics::Metrics;
 #[cfg(test)]
 pub(crate) use operations::compute_execution_results_checksum;
-pub use operations::execute_finalized_block;
 use operations::speculatively_execute;
 pub(crate) use types::{
     BlockAndExecutionArtifacts, ExecutionArtifact, ExecutionPreState, SpeculativeExecutionResult,
